+++ src/components/containers/friends/children/friendList/FriendList.js
