+++ src/components/routes/Routes.js
@@ -7,9 +7,9 @@ import ShareTrack from "../containers/shareTracks/ShareTrack";
 import DownloadTrack from "../containers/downloadTracks/DownloadTrack";
 import Welcome from "../containers/welcome/Welcome";
 import Login from "../containers/login/Login";
-import Friends from "../containers/friends/Friends";
 import VisualizeNotifications from "../containers/visualizeNotifications/VisualizeNotifications";
 import { LoggedIn, LoggedOut } from "@solid/react";
+import Friends from "../containers/friends/Friends";
 
 /**
  * It is in charge of the different routes of the application
